requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
attrs==25.3.0
babel==2.17.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
commonmark==0.9.1
contourpy==1.3.2
cycler==0.12.1
daal==2025.5.0
decorator==5.2.1
einops==0.8.1
faicons==0.2.2
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.5.1
great-tables==0.17.0
h5py==3.13.0
htmltools==0.6.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mne==1.9.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
polars==1.30.0
pooch==1.8.2
propcache==0.3.1
psutil==7.0.0
pyaml==25.5.0
pycparser==2.22
pymer4==0.8.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.32.3
rpy2==3.6.0
rpy2-rinterface==3.6.0
rpy2-robjects==3.6.0
scikit-learn==1.6.1
scikit-learn-intelex==2025.5.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
statsmodels==0.14.4
sympy==1.14.0
tbb==2022.1.0
tcmlib==1.3.0
threadpoolctl==3.6.0
torch==2.7.0
torch-geometric==2.6.1
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarl==1.20.0
zipp==3.22.0
